requirements.txt
Django==6.0.3
djangorestframework==3.16.1
psycopg[binary]==3.3.3
python-dotenv==1.2.2
Pillow==10.4.0
djangorestframework-simplejwt==5.5.1
channels==4.1.0
channels-redis==4.2.0
daphne>=4,<5
redis==6.4.0
firebase-admin==7.1.0
google-cloud-translate==3.25.0
